fully replace old events cache with new events cache
