+++ src/ctc/evm/event_utils/__init__.py
@@ -1,1 +1,2 @@
-from .event_crud import is_event_hash, async_get_events
+from .event_crud import async_get_events
+from .event_metadata import async_get_event_timestamps
