+++ src/ctc/evm/event_utils/event_legacy_utils.py
