--- src/ctc/evm/event_utils_new/__init__.py
@@ -1,2 +1,0 @@
-from .event_crud import async_get_events
-from .event_metadata import async_get_event_timestamps
