+++ src/ctc/evm/event_utils_old/__init__.py
@@ -1,0 +1,1 @@
+from .event_crud import is_event_hash, async_get_events
